+++ src/ts/Core/Models/ABGroup.ts
@@ -44,4 +44,5 @@ export const FakeDisabledABTest = new DisabledABTest(16, 17);
 // Add actual A/B tests below
 
 export const AuctionV5Test = new DisabledABTest(15, 16, 17);
+export const VastParsingStrictTest = new DisabledABTest(); // Testing new vast parser logic
 export const ByteDanceCTATest = new ABTest(17);